Add ApodiniDocumentExport dependency to refactor ApodiniMigratorShared
