+++ Package.swift
@@ -36,7 +36,9 @@ let package = Package(
         .package(url: "https://github.com/apple/swift-collections.git", .upToNextMajor(from: "1.0.0")),
 
         // testing runtime crashes
-        .package(url: "https://github.com/norio-nomura/XCTAssertCrash.git", from: "0.2.0")
+        .package(url: "https://github.com/norio-nomura/XCTAssertCrash.git", from: "0.2.0"),
+        
+        .package(name: "ApodiniDocumentExport", path: "../ApodiniDocumentExport")
     ],
     targets: [
         // The lowest level ApodiniMigrator package providing common API used across several targets, including
@@ -46,7 +48,8 @@ let package = Package(
             dependencies: [
                 .product(name: "PathKit", package: "PathKit"),
                 .product(name: "FineJSON", package: "FineJSON"),
-                .product(name: "Yams", package: "Yams")
+                .product(name: "Yams", package: "Yams"),
+                .product(name: "ApodiniDocumentExport", package: "ApodiniDocumentExport")
             ]
         ),
 
@@ -97,7 +100,8 @@ let package = Package(
                 .target(name: "ApodiniMigrator"),
                 .target(name: "ApodiniMigratorCompare"),
                 .target(name: "ApodiniMigratorClientSupport"),
-                .product(name: "Logging", package: "swift-log")
+                .product(name: "Logging", package: "swift-log"),
+                .product(name: "ApodiniDocumentExport", package: "ApodiniDocumentExport")
             ],
             resources: [
                 .process("Resources")
@@ -112,7 +116,8 @@ let package = Package(
             dependencies: [
                 .target(name: "RESTMigrator"),
                 .product(name: "ArgumentParser", package: "swift-argument-parser"),
-                .product(name: "Logging", package: "swift-log")
+                .product(name: "Logging", package: "swift-log"),
+                .product(name: "ApodiniDocumentExport", package: "ApodiniDocumentExport")
             ]
         ),
 
@@ -124,7 +129,8 @@ let package = Package(
                 "RESTMigrator",
                 "ApodiniMigratorCompare",
                 "ApodiniMigratorClientSupport",
-                .product(name: "XCTAssertCrash", package: "XCTAssertCrash", condition: .when(platforms: [.macOS]))
+                .product(name: "XCTAssertCrash", package: "XCTAssertCrash", condition: .when(platforms: [.macOS])),
+                .product(name: "ApodiniDocumentExport", package: "ApodiniDocumentExport")
             ],
             resources: [
                 .process("Resources")
