+++ Sources/ApodiniMigratorShared/Extensions/KeyedEncodingContainerProtocol+Extensions.swift
@@ -1,0 +1,17 @@
+//
+// This source file is part of the Apodini open source project
+//
+// SPDX-FileCopyrightText: 2019-2021 Paul Schmiedmayer and the Apodini project authors (see CONTRIBUTORS.md) <paul.schmiedmayer@tum.de>
+//
+// SPDX-License-Identifier: MIT
+//
+
+// MARK: - KeyedEncodingContainerProtocol
+extension KeyedEncodingContainerProtocol {
+    /// Only encodes the value if the collection is not empty
+    public mutating func encodeIfNotEmpty<T: Encodable>(_ value: T, forKey key: Key) throws where T: Collection, T.Element: Encodable {
+        if !value.isEmpty {
+            try encode(value, forKey: key)
+        }
+    }
+}
